requirements.txt
apscheduler>=3.10.4
feedparser>=6.0.11
httpx>=0.27.0
pydantic>=2.7.0
PyYAML>=6.0.1

